+++ examples/plugin-workspace/.ontocore/plugins/demo_providers.sh
@@ -1,0 +1,1 @@
+../../demo_providers.sh